Use 32 bit cookie

+++ pkg/network/event_common.go
@@ -220,7 +220,7 @@ func (s StatCounters) IsZero() bool {
 	return s == StatCounters{}
 }
 
-type StatCookie = uint64
+type StatCookie = uint32
 
 // ConnectionStats stores statistics for a single connection.  Field order in the struct should be 8-byte aligned
 type ConnectionStats struct {
